t Rename samplewidgets.h and .cpp, in preparation for adding other ways of using the .ui file
