+++ examples/accessing_widgets/accessing_widgets_lib/samplewidgets01.h
